+++ src/collective/behavior/seo/browser/metarobots.py
@@ -30,13 +30,4 @@ class MetaRobotsViewlet(common.ViewletBase):
         # If there is no restriction, we explicity allow indexing
         if not content:
             return 'all'
-        content = list(content)
-        # For unavailable_after we need to inform the expiration date
-        if 'unavailable_after' in content:
-            content.remove('unavailable_after')
-            expiration = self.context.expires
-            if expiration:
-                content.append(
-                    'unavailable_after: {:%d %b %Y %H:%M:%S %Z}'.format(expiration)
-                )
-        return u', '.join(content)
+        return content